# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: RealEstate AI API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /properties:
    get:
      operationId: listProperties
      tags: [properties]
      summary: List user's properties
      responses:
        "200":
          description: List of properties
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Property"
    post:
      operationId: createProperty
      tags: [properties]
      summary: Create a new property
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PropertyInput"
      responses:
        "201":
          description: Created property
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
...
  /properties/{id}:
    get:
      operationId: getProperty
      tags: [properties]
      summary: Get a property by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Property
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
        "404":
          description: Not found
    patch:
      operationId: updateProperty
      tags: [properties]
      summary: Update a property
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PropertyUpdate"
      responses:
        "200":
          description: Updated property
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
    delete:
      operationId: deleteProperty
      tags: [properties]
      summary: Delete a property
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Projects ───────────────────────────────────────────────────────────────
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List user's development projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
...
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Zones ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Property:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        address:
          type: string
        city:
          type: string
        neighborhood:
          type: ["string", "null"]
        propertyType:
          type: string
          enum: [casa, depa, local, terreno]
        status:
          type: string
          enum: [activa, vendida, en_rehab]
        purchasePrice:
          type: number
        currentValue:
          type: ["number", "null"]
        rentAmount:
          type: ["number", "null"]
        expensesMonthly:
          type: ["number", "null"]
        netIncome:
          type: ["number", "null"]
        occupancyRate:
          type: ["number", "null"]
        nextInspectionDate:
          type: ["string", "null"]
        images:
          type: array
          items:
            type: string
        isPublic:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, userId, title, address, city, propertyType, status, purchasePrice, images, isPublic, createdAt]
    PropertyInput:
      type: object
      properties:
        title:
          type: string
        address:
          type: string
        city:
          type: string
        neighborhood:
          type: string
        propertyType:
          type: string
          enum: [casa, depa, local, terreno]
        status:
          type: string
          enum: [activa, vendida, en_rehab]
        purchasePrice:
          type: number
        currentValue:
          type: number
        rentAmount:
          type: number
        expensesMonthly:
          type: number
        netIncome:
          type: number
        occupancyRate:
          type: number
        nextInspectionDate:
          type: string
        images:
          type: array
          items:
            type: string
        isPublic:
          type: boolean
      required: [title, address, city, propertyType, status, purchasePrice]
    PropertyUpdate:
      type: object
      properties:
        title:
          type: string
        address:
          type: string
        city:
          type: string
        neighborhood:
          type: string
        propertyType:
          type: string
          enum: [casa, depa, local, terreno]
        status:
          type: string
          enum: [activa, vendida, en_rehab]
        purchasePrice:
          type: number
        currentValue:
          type: number
        rentAmount:
          type: number
        expensesMonthly:
          type: number
        netIncome:
          type: number
        occupancyRate:
          type: number
        nextInspectionDate:
          type: string
        images:
          type: array
          items:
            type: string
        isPublic:
          type: boolean
...
    ProjectPhase:
      type: object
      properties:
        name:
          type: string
        progress:
          type: integer
        status:
          type: string
          enum: [pending, active, completed]
      required: [name, progress, status]
    Project:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        location:
          type: string
        description:
          type: ["string", "null"]
        type:
          type: ["string", "null"]
        status:
          type: string
          enum: [preventa, construccion, entrega, completado]
        roiEstimated:
          type: ["number", "null"]
        totalInvestment:
          type: ["number", "null"]
        raisedAmount:
          type: ["number", "null"]
        phases:
          type: array
          items:
            $ref: "#/components/schemas/ProjectPhase"
        isPublic:
          type: boolean
        images:
          type: array
          items:
            type: string
        investorCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, userId, title, location, status, phases, isPublic, images, investorCount, createdAt]
    ProjectInput:
      type: object
      properties:
        title:
          type: string
        location:
          type: string
        description:
          type: string
        type:
          type: string
        status:
          type: string
          enum: [preventa, construccion, entrega, completado]
        roiEstimated:
          type: number
        totalInvestment:
          type: number
        raisedAmount:
          type: number
        phases:
          type: array
          items:
            $ref: "#/components/schemas/ProjectPhase"
        isPublic:
          type: boolean
        images:
          type: array
          items:
            type: string
      required: [title, location, status]
    ProjectUpdate:
      type: object
      properties:
        title:
          type: string
        location:
          type: string
        description:
          type: string
        type:
          type: string
        status:
          type: string
          enum: [preventa, construccion, entrega, completado]
        roiEstimated:
          type: number
        totalInvestment:
          type: number
        raisedAmount:
          type: number
        phases:
          type: array
          items:
            $ref: "#/components/schemas/ProjectPhase"
        isPublic:
          type: boolean
        images:
          type: array
          items:
            type: string

    # ── Zones ─────────────────────────────────────────────────────────────────
...