# lib/api-zod/src/generated/types/property.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RealEstate AI API
 * OpenAPI spec version: 0.1.0
 */
import type { PropertyPropertyType } from './propertyPropertyType';
import type { PropertyStatus } from './propertyStatus';

export interface Property {
  id: number;
  userId: number;
  title: string;
  address: string;
  city: string;
  /** @nullable */
  neighborhood?: string | null;
  propertyType: PropertyPropertyType;
  status: PropertyStatus;
  purchasePrice: number;
  /** @nullable */
  currentValue?: number | null;
  /** @nullable */
  rentAmount?: number | null;
  /** @nullable */
  expensesMonthly?: number | null;
  /** @nullable */
  netIncome?: number | null;
  /** @nullable */
  occupancyRate?: number | null;
  /** @nullable */
  nextInspectionDate?: string | null;
  images: string[];
  isPublic: boolean;
  createdAt: Date;
}