# lib/api-zod/src/generated/types/propertyUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RealEstate AI API
 * OpenAPI spec version: 0.1.0
 */
import type { PropertyUpdatePropertyType } from './propertyUpdatePropertyType';
import type { PropertyUpdateStatus } from './propertyUpdateStatus';

export interface PropertyUpdate {
  title?: string;
  address?: string;
  city?: string;
  neighborhood?: string;
  propertyType?: PropertyUpdatePropertyType;
  status?: PropertyUpdateStatus;
  purchasePrice?: number;
  currentValue?: number;
  rentAmount?: number;
  expensesMonthly?: number;
  netIncome?: number;
  occupancyRate?: number;
  nextInspectionDate?: string;
  images?: string[];
  isPublic?: boolean;
}